mediasite/requirements.txt
Django==1.7
Pillow==2.5.3
argparse==1.2.1
django-bootstrap-form==3.1
django-suit-ckeditor==0.0.2
simplejson==3.6.3
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
